feat: implementa a passagem de telas inicial
