+++ .gitignore
@@ -1,2 +1,3 @@
 /View/nbproject/private/
 /View/build/
+/View/dist/
