+++ View/src/Model/DAO/PublicAgentDAO.java
@@ -9,6 +9,6 @@ public interface PublicAgentDAO {
     void update(String phoneNumber1, String address, String email, Integer idPublicAgent);
     PublicAgent findById(Integer idPublicAgent);
     List<PublicAgent> findAll();
-    boolean containsUser(String user, String password);
+    PublicAgent containsUser(String user, String password);
     void disable(Integer idPublicAgent);
 }
